RFC: Trace Identifiers (#109)
The Opentracing model of computation specifies two primary object types, Spans and Traces, but does not specify identifiers for these objects. The lack of identifiers makes it very difficult to correlate tracing data with data in other systems. This complicates a number of important tasks, and prevents the creation of reusable trace observers.

To address these difficulties, this draft proposal extends the OpenTracing SpanContext interface to include SpanID and TraceID accessors.